requirements.txt
fugashi[unidic-lite]
python-dotenv
ibm-watson
ibm-cloud-sdk-core
jiwer
tqdm